requirements.txt
torch
torchvision
pillow
scikit-learn
